+++ src/package/themes/default/index.ts
@@ -2,6 +2,7 @@ import base from './base';
 import button from './components/button';
 import card from './components/card';
 import input from './components/input';
+import numberInput from './components/number-input';
 import formField from './components/form-field';
 import textarea from './components/textarea';
 import link from './components/link';
@@ -37,6 +38,7 @@ export default {
         button,
         card,
         input,
+        numberInput,
         formField,
         textarea,
         link,